move generic functions to core class
